implemented computer container list, removed OnDetach and repurposing OnAttach
Using this implementation of @Computer for an attached computer list means it is one less thing for developers to maintain in their peripheral code, they can just simply add this annotated field and it will always contain an up-to-date list of attached computers. as such the OnDetach annotation was removed as this no longer has a need. the OnAttach annotation will be repurposed tomorrow to construct and mount the supplied IMounts to the computers, this removes yet another thing that developers must maintain by the Peripheral-Framework automagically mounting and unmounting IMounts.