package restructure and new virtual filesystem mounting
virtual filesystem mounting is now completely controlled by the Peripheral-Framework. You are required to specify the indended class(es) that you wish to be mounted via the LuaPeripheral annotation. When the first instance of a peripheral is attached to a computer the Peripheral-Framework will instansiate and mount the specified mount class(es). When the last instance of a peripheral is detached from a computer the Peripheral-Framework will unmount the mounts; due to a bug in ComputerCraft, if the mounting peripheral is removed from the computer the mount will also be removed, there is no fix for this normally, and the Peripheral-Framework is unable to mitigate this bug, bug info: http://www.computercraft.info/forums2/index.php?/topic/19768-